server/pyproject.toml
[tool.poetry]
name = "braine_backend"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langchain-core = "^0.3.15"
langchain-community = "^0.3.5"
langchain-groq = "^0.2.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
nameko = "3.0.0-rc11"
faiss-cpu = "^1.9.0"
pypdf2 = "^3.0.1"
sqlmodel = "^0.0.22"
sqlmodel-controller = "^0.1.8"
openai = "^1.54.3"
tiktoken = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
